refactored inputs, changed exception, etc
